Bring over code needed for the steps
This code will handle the steps progressing (`update_and_advance`) making use of form objects and decision trees.

Also we update the database to add a couple of new columns, one is the `navigation_stack` that will hold the different steps a user navigates and the other is the `kind` attribute, that will be needed in a follow-up commit.