Separate the Caution and Conviction journeys
We will have a separate decision tree for each of these kind of checks.

This will make it easier to map out the different journeys and avoid having a very long decision tree.

Move the `caution_date` step from the current `Check` namespace to the new `Caution` namespace and update decision trees.