Use (or rather attempt to) GOV.UK Frontend
This is a first attempt and kind of an experiment at trying to replace our current `govuk_elements_rails` and `govuk_template` gems, with the new design system pulled as a node module.

Refer to README to know how to  install it.

As all class names, and quite a lot of the markup has changed dramatically, we will need to adapt many things.

In the end it might be just too much work (specifically with regards to the `govuk_elements_form_builder` gem) and not possible to accomplish it at the moment due to time and costs contrainsts.