Migrate `radio_button_fieldset` to new design system
For now, we are taking the approach of migrating one by one the absolute minimum of form elements that we will require in this service.

In order to do this, we reuse as much as possible from the original gem, and override/reimplement the methods that are no longer possible to use due to producing old or invalid markup and hardcoded old style class names.

Please note the summary error in the top of the page it is not migrated yet.