Migrate the error summary to new design system
In a similar way to what we did with the radio options, we are only overriding the methods from the gem that are necessary to produce the new markup.

Also, we have to separate the `back` link from the error summary as, in the new design system, these 2 components are in different places in the page html hierarchy.