Use new version of `gov_uk_date_fields` gem
For now we are referencing the branch, until a new release of the gem is produced.

The gem has been updated to produce new markup and styles according to the new design system.

This will serve as a proof of concept to see if the updates to the gem serves the purpose for our service.