Fix missing assets
Some assets (fonts and images) were not found as they were using a digest after the precompile, butthere are some hardcoded references in the `govuk-frontend` files that will not be able to use the rails digest mechanism.

In order to fix this problem without having to modify the `govuk-frontend` files to embed rails code (like `asset_path()`) we copy over the non-digested files along with the digested ones.

Also removed the ENV variable `SECRET_KEY_BASE` as Rails 5 uses a different mechanism and we don't need it anymore.