Mark a check as completed
https://www.pivotaltracker.com/story/show/164861902

In order to know when a check has been completed, we change their `status` attribute from `in_progress` to `completed`.

This also allow us to handle errors when trying to modify a completed check, and in the future will let us know which checks we can purge from the database, among other functionality we might want to introduce.

Note: the copy in the views is temporary until we have the final sign-off.