Add missing error message and remove unneccessary HTML page
There was a missing error message on the condition_complied page. I've added one in that is more appropriate than the filler one that was being put in, but these will still need to be looked at by content and updated accordingly, in due course. I've also removed an unnecessary HTML file that Jesus spotted I'd worked on previously.