Serve static assets with nginx, locally (docker-compose)
This is a first step towards serving the static assets with nginx instead of Rails, and also eventually the "maintenance" page.

Adapted the `docker-compose` configuration to replicate what would be a real production setup, with nginx reverse proxy, and created a basic Dockerfile.nginx to create the nginx image.

Added to CircleCI configuration the build and push of the nginx docker image (but it will not have any effect yet on k8s until we also create the corresponding configuration).

Removed static error pages in /public (except one that will be used by nginx) as any other kind of errors are served by Rails directly.