Make errors specific for each form
Now we are re-using some DB attributes in different form objects, we need a way to specify different error messages depending on the step we are.

I didn't remember, but the gem provides this mechanism out of the box. The only thing we need to do is reorganise a bit the locales, as shown in this PR.

We should do the same for the rest of errors.