Unify the convictions types
A single value-object to unify all the different conviction types, with ability to get the parent or the children, and this way, have a tree of convictions that can be easily traversed.

The main objective of this refactor will be to get rid of the individual conviction DB fields existing at the moment (community_order, custodial_sentence, discharge, etc.) and associated form objects, controllers and hopefully even views (but probably we will need to use partials to customise some copy).

This is still WIP but will be pushing the changes as I go, up to a point where we decide if this is worth doing or not (because maybe, we find out this will not simplify but instead make the code harder to read and follow).