Fix test coverage
As we are not yet using all convictions, or we've stopped using the old value-object, the test coverage is complaining.

For now we will mark as `:nocov:` this code, but this will be fixed itself as we implement what is missing and do a huge cleanup of the old code and value-object.