Adapt form_builder to support virtual attrs
As we will be reusing one single form object and view for the conviction subtypes, and storing the selection to the same DB attribute (`conviction_subtype`), we can't customise the legend of the radio `radio_button_fieldset`, because with the current code, the locales lookup goes to `conviction_subtype` always.

The solution (quite easy actually) is to support passing a "virtual attribute", so this can be parameterized depending on the main conviction type. So for example, the "real attribute" is still `conviction_subtype` but the virtual attribute when we are showing the list of community orders, will be `community_order` so it looks for the appropiated locale.