Remove unused convictions code and DB fields
This is the final PR to do the cleanup of the unused conviction types, and their database fields.

I've left the `hospital_order` although the prototype doesn't have it, as an example of a conviction that goes direct to the date input (doesn't have children).

Once the prototype is updated, and depending where this conviction goes, we can move it or delete it.