Split expiry calculator into two
As the calculations are fundamentally different for caution and convictions, it will be better to have 2 separate calculators, that can be implemented in different ways.

As a consequence of doing this, now we can get rid of some code we no longer need in the `DisclosureCheck` model, for example.

Reorganise a bit the specs.

Note: this PR doesn't incude the results page for convictions, nor actually any calculations. This is just a small code reorganisation to prepare for the future.