Delegate some frequently used methods
These methods are used in several views through the `@form_object` instance, to customise some of the locales.

Instead of delegating them in each of the form objects, we move them to the superclass.

If they are only used in one single form object, like `conviction_length_type` then do not moved.