Make the `known_date` header dynamic
Depending on the subtype of conviction, the header of this step can vary, so for example in most cases we talk about an `order` but in a few cases it can be a `sentence`, a `discharge` or even a `treatment`.

In order to do that we've bumped the version of the `gov_uk_date_fields` in this PR https://github.com/ministryofjustice/gov_uk_date_fields/pull/36

Updated the cucumber tests to also test properly these headers.

Please note this is the first PR, but another one will be needed to do the same with other steps (and update the FormBuilder to also use the `i18n_attribute` for consistency).