Update views and cukes
We add the subsection in the views, and also created a new step definition to test for this subsection, but we only do this in the discharge as an example, there is no need to go and add this test to all cukes as it is implicitly tested when selecting the each of the convictions.

The file previously named `conviction_financial_discharge` has been renamed to `conviction_discharge`.