Rename service and reorg locales
The name of the service has changed.

I've done a bit of reorg in the locales, so we don't have many 1 string files.
Also removed the `home` page_title locale as it is not used.

Note: fixed as part of this PR, although unrelated, the cookies page as it was missing a closing div.