Refactor some calculators and tests
We needed another calculator for:

```"If a length (B) was given: A + B
If no length was given: A + 24 months"```

So in order to organise a bit better the current calculators and the new one, an `AdditionCalculator` has been added that contains 3 types of sub-calculators that cope with most possibilities.

Also the calculator tests has been moved to `conviction_type_spec` where we now test all the attributes of the value object.