Reset `caution_type` dependent attributes
If CautionTypeForm changes with regards to the database (for example, if user advance but then go back and change their answers) we need to ensure some dependent attributes down the line are also reset, to avoid inconsistencies in the decision trees or result page.

This is easily done returning early from the `#persist!` method if there are no changes in the form.

I will do a similar thing with convictions in a separate PR.

Note: unrelated, but I removed the safe navigation operator in the `#under_age?`, ensuring we do the validations if we have a `disclosure_check`.