Hide armed forces convictions
We will not have these for MVP, but as we've done most of the work and we will eventually need them, instead of removing everything let's just hide the parent group.

This mechanism can also be used in the future to show/hide any other convictions if needed.

Added a default cucumber profile that skips scenarios tagged as `@skip`.