Combine conviction length details in results page
The conviction length is split into two bits of information, the actual length (a number) and the metric (weeks, months, years or `no_length`).

Previously we had these separate in the summary table in the results page, but as part of the service review, it was decided to combine it into one line, if possible.

We do this by extending the functionality of the presenters, to be able to override the value of an attribute and use a custom method to build the one-liner.

Also, I've removed the individual test of the `#question_attributes` method, because it is tested more in-deep in the `describe '#summary'`, so it was redundant. This method is now private.