Use a non-root nginx container
In order to replicate as much as possible the setup we have on kubernetes, where we can't run root containers, we are going to change the nginx image to the `bitnami` one and adapt the config so it works.

A similar approach to this, will be needed in our staging kubernetes environment.