Add GA custom dimension (spent)
Add to the completed transaction an extra bit of information whether the caution or the conviction is spent (yes/no values).
This is done with custom dimensions.

A bit of refactor and moving code related to GA to a separate helper `analytics_helper.rb`

Removed the unused method `#display_result`

Fixed the results page title.