Filter query string from GA outbound links
The links to surveymonkey will contain query params with some variable information.

This information, at the moment, is not personal nor secret but will generate many different URLs that will make it difficult to analyse in GA.

Also in the future other ourbound links might contain private data, so better to filter this out.