Bypass `under_age` step if adults journey is enabled
This code is very simple in essence, we just check if we have the "adults" cookie, and if so, we let the user continue to the adults journey after the `under_age` step.

But as the decision trees doesn't have access to cookies nor controller-specific code, we must workaround this limitation.