Prepare conviction types for adults
Some changes to the form object, in the same way we did for cautions.

Separate parent types for youth and adults in the value object.

In essence, all new parent types, and their children, must have different names (even if they are identical to the youth ones) so we can customise the copy, they can have different calculators, and they track differently in GA, etc.