Add embedded survey and update phase link
We now have the "real" surveys.

Add the end of service embedded survey (like we did in C100) and update the link in the phase banner.

Ensure we propagate `check=unknown` to the survey if we don't have a current_disclosure_check record, or we have it, but it doesn't have the `kind` attribute yet.