Refactor detention and prison calculators
As most of the code is common between them, we take a similar approach to the `AdditionCalculator` and create subclasses that just customise a bit the rehabilitation periods.

Another needed functionality is the requirement to validate the upper limit (only for the `SuspendedPrison`).

For now this is prepared but as part of a separate PR the validation will be introduced in the form object.

At the moment it will just raise an exception if the calculation does not meet the requirements. In practice, this exception should never happen as long as we validate the input in the form objects.