Improve prevention/reparation value-object naming
Just an internal renaming, for consistency and to avoid confusion, not affecting user interface.

The previous name apart from being too long, gave the impression it was an order in itself, instead of a "group" of orders, or category (the `and` was changed to `or` in the user interface as per PR #197).