Add details to date steps
Add a `<details>` block to all date steps we have in the application, to explain more about it.

All are the same, except for conditional caution, where the copy is slighly different.

As we are reusing the same markup, it has been extracted to a partial that can be passed a few variables.