Use view variants
Instead of having many if-else or complicated locales hierarchy, we will be rendering a different `results/show` view (variant) depending of the type of check (caution / conviction) and if it is spent or not.

This way, the copy that is different will be self-contained in the view itself, and the copy that is common will go into partials that are shared between all the different variants.

There are 5 different view variants:

* caution_spent
* caution_not_spent
* conviction_spent
* conviction_not_spent
* conviction_never_spent