Add landing page for MVP info and opt-out
We create a new MVP task with a single step for now, named `info`, which will be the landing page to hold information about the MVP and a mechanism to opt-out.

This is still WIP so some assumptions have been made, also we are waiting for the copy and the list of references, as we will want to do a whitelist.

The decision tree at the moment takes the participant to the home page, but this will change soon and there will be a confirmation page after they save their opt-out status.

A small change to the routes, to support URLs in the form: `steps/mvp/info/:id` where `:id` is the reference of the participant YOT.