Whitelist MVP participant references
Created a rake task to generate a variable number of unique, short, reference codes (I used this to generate the 180 we've added to the YOT contact list).

Also, implemented a simple mechanism to whitelist the reference code provided in the URL. Just a simple SHA256 digest of the reference, and we store the list of digests to check if the given reference exists or not.

If we ever need to add more participants, it is just a matter of generating a new reference code (checking first it doesn't exist, must be unique) and get the SHA to add to the `participants.yml` file.