Enable credentials for MVP participants
Each MVP participant will have a unique set of credentials, these being their reference code.
User and password, for simplicity will be the same.

Still we are going to support a separate set of credentials for internal use. These credentials will be exposed in the ENV.

Note: I've removed the production 'test' reference from the participants.yml file, and replaced with another one (that will be shared internally for the team).