Enable GA userId for MVP pilot
At least during the pilot, we will need a way of measuring success and we will need to know how many checks have been completed per YOT.

This will link together the analytics belonging to each YOT so it is easier to evaluate the metrics we get.

There is no personal identification involved.