Refactor some conviction methods into decorator
The `ConvictionType` value-object was getting too big and it might grow further with motoring convictions.

A new module, that is included in the value object, named `ConvictionDecorator` will group together some of these utility and query methods to reduce complexity when testing and also avoid having more attributes than neccessary (we still maintain `skip_length` because this one is used by a lot of different convictions, not just 2 like compensation).

Also, added to the superclass `ValueObject` an `inquiry` method to simplify doing things like checking if a conviction is of a specific type or not.

So for example, it is now possible to do this (might be useful for decision trees):

```ruby
c = ConvictionType::ADULT_MOTORING_FINE

c.inquiry.foobar?
=> false
c.inquiry.adult_motoring_fine?
=> true
c.parent.inquiry.adult_motoring?
=> true
```