Refactor completion mechanism
Now we have 2 different completions: one for the individual check (a caution or a conviction) and another for the whole set of checks (one or more) past the CYA page.

In order to be more precise when a check is completed, the action of completing an individual check has been moved to the decision tree who is responsable for knowing when to redirect the user to the results page.

In following PRs we will introduce a proper error handling when an individual check is completed or when the whole report is completed, along with the functionality to cancel an "in-progress" check.