Do not mark report as completed on CYA
As we are using a little hack to feature-flag the multiples work, while it is under development, there was a side effect to this, where the report would also be marked as completed the moment we hit the `results` controller (even if it was just to redirect to CYA).

Reorganised the callbacks so the redirect takes precedence and halts any other callbacks (in particular the one marking the report as completed).