Extend error handling to deal with reports
Same as we had with check completed, we now have a new scenario where a "report" (definition of report being a bunch of cautions and conviction checks) have been completed.

This happens at the very end of the service once the loop over the multiple caution or convictions have been done and the user wants to see the final results.

This should be transparent to a user without the 'multiples' feature flag enabled.

Some copy is still TBD.