Show link to go back to check your answers
After the first check has been completed, following ones will show a 'Go back to check your answers' link below the 'Continue' button.

This, in essence, will take the user to the CYA and leave the check in the session, but still marked as `in-progress`, so the CYA page must cope with this (the presenter must exclude not completed checks).