Fix back links after adding new checks
With the introduction of the CYA page, when adding a new caution or conviction, the back link in the first step will point to the home page, instead of the CYA as it would be expected.

Also, more importantly, if the user clicks this `back` link and loads the home page, the session will be reset.

This is fixed so the `back` link after adding a new caution or conviction takes the user back to the CYA instead of the home page, without overriding any existing navigation stack.