Add link to CYA on compensation not paid
If multiples is enabled and we have at least one completed check, we will let the user jump to CYA page if they fall into this kickout page, otherwise they don't have a way out.

Also added a 'start again' button for consistency with other pages. This will reset the session.

Added the `govuk-link--no-visited-state` class to a few other links.