Default age of new convictions inside a group
When adding a conviction to an existing proceeding (group), we can skip the 'under_age' question, as all checks inside the same group will have the same age.

We pick the first check in the group and take this age to use it in the new check being initialized.

Also, as a consequence, we redirect to the question after (conviction type).