Implement new methods in module
These new methods will serve a purpose later on (in particular, the one named `#offence_type`) as will allow us to generalise a few classes as this method will return the correct caution type or conviction subtype depending on the kind of the disclosure check.

Moved the class method `.find_constant` to the `ValueObject` superclass, as all value-objects should have the same interface.