Refactor the caution calculator
The new calculator performs the same as the previous (conditional) one but also handles non-conditional (simple) cautions.

Exposed a `#calculator_class` method in the `CautionType` value-object for consistency with ConvictionType so now any kind of offence offers the same method to retrieve its calculator.

This will have a benefit in next commit where we will be able to remove a lot of specific logic in favor of generic code.