Preliminary PoC for multiples calculator
This is a proof of concept, and a preliminary work for what could be the multiples calculator.

It is still WIP and not properly working under all scenarios, but it currently will work for:

- A group of convictions in the same proceedings (any type of conviction, but with the limitation at the moment of at most 1 prison sentence in the group).

- Any number of individual (separate proceedings) cautions (simple or conditional) and, at most, 1 conviction of any type (if there is more than 1 conviction, the calculation will be wrong).

The way to use it (at least initially, as I expect some refactors) is, somewhere in the Results presenter to call:

```ruby
calculator = Calculators::Multiples::MultipleOffensesCalculator.new(current_disclosure_report)
calculator.process!
```

And then, in each group where the spent date need to be shown, call the following to retrieve the date (or `:never_spent` or `:no_record`):

```ruby
calculator.spent_date_for(check_group)
```

Where `check_group` is the CheckGroup instance.