Show results page for multiple checks
At the moment just a basic show page, re-using the `CheckAnswersPresenter` to list again all the cautions and convictions.

Next steps, as follow-up PRs, will be to introduce the spent date panels in each of the groups, making use of the `MultipleOffensesCalculator` from PR #281.

Also, adding the remaining copy in the results page as per design.