Purge in-progress checks on report completion
Once we reach this point, there is not going back, and we don't need to maintain around the in-progress checks, because they don't serve any purpose.

The advantage of doing this, is we don't need to remember anymore in the results page to use the scope `#completed` in the disclosure checks collection, which simplify a bit the code.