Retrieve the spent date from the calculator
For each of the different groups in the results page, we retrieve the spent date from the `multiple_offenses_calculator` and build little `SpentDatePanel` instances to hold the logic to know what to show in the results page.

Still to do in following PR(s): remaining copy in the results page, specially, the header must be dynamic depending if we have or not spent convictions.