Simplify results summary tables
As per Simon new designs for multiples, we are going to partially implement this for the current version of the service, as it makes for a simplified and more easy to understand summary table.

Essentially we remove 1 row in cautions, and 2 rows in convictions.

There is still a change that we might want to do, which is replacing the word `conviction` with `sentence` in the tables, but only for certain conviction types.
This will require some UR so it is not implemented yet.

Adapt the tests and locales.