Rename value of conviction type for consistency
We renamed `Community or youth rehabilitation order (YRO)` to `Referral or youth rehabilitation order (YRO)` but only on the user interface.

For consistency, and because it will have an effect on analytics and metrics, we should also get rid of the word `community` in the value object.

As per Simon suggestion we can internally rename `COMMUNITY_ORDER` type to `REFERRAL_SUPERVISION_YRO` to be more accurate.

Better to do it now before we start getting checks for this type.