Show armed forces, was hidden for YOTs
This was already implemented into the service but hidden as for the YOT pilot there was no audience to test with.

Added the hint text, reordered alphabetically and fixed a couple of calculations.

This can be deployed to staging for testing, make any further amendments and, if we decide so, release it.
If not, we can use again the `DISABLED_PARENT_TYPES` array to hide it but leave it prepared for a quick release at a later date.