Enable postgres `verify-full` SSL mode
For production environments we are going to force the verify-full SSL mode in all connections to the database.

It will be overriden (`DATABASE_SSLMODE=disable`) when using docker-compose so we are still able to run and test the application locally.

This follows best practices and guidance from Cloud Platforms here:

https://user-guide.cloud-platform.service.justice.gov.uk/tasks.html#ssl-connections-with-rds