Add days spent on bail step
As part of the bail journey, this is the follow-up question after the previous yes-no question.

It is not yet visible in the "normal" journey but can be accessed directly knowing the URL.

Fixed an issue in our custom form builder where a `text_field` would not maintain their value between reloads.