Implement check boxes in custom form builder
Re-implement `check_box_fieldset` to override the original `GovukElementsFormBuilder::FormBuilder` method and produce the new design system markup.

As we did with other methods, this is kind of a monkey patch and we only re-implement the absolute neccessary in order to produce the new markup, without having to rewrite the whole gem, and eventually, hopefully, we will be able to replace all this code with a new gem.