Update date controllers
As we are going to be adding a new attribute to each of these forms, it will be better to use the `#build` method that already takes any attributes in the form objects, instead of having to pass individually each of the attributes.

Functionally, there is no difference between one form or the other. It just saves some coding.