Remove the pilot landing page (redirects to home)
We do not need the pilot landing page anymore, but some users might still use it or have it on bookmarks or emails, so we are going to maintain the route but redirect to the homepage.

That way we can still track these visits.

Also, some minor copy amendments to the home page as per content designer suggestions.