Change bail offset from half a day to a whole day
Ticket: https://mojdigital.teamwork.com/#/tasks/19868485

This is an urgent change as per sentencing policy:
Courts already calculate days on bail into the sentence, to keep calculations accurate we need to have a whole day taken off the spent date for each day spent on bail tag.

Instead of half a day as before, now the bail time substract the total amount (of bail time) from the sentence which effectively is equivalent to multiply by -1.