Fix the order of bail time calculations
Note: bail is not yet "visible" in production. This can be tested on dev/staging environments.

Previously, we deducted the bail time before we looked at the rehabilitation periods.

This is incorrect - the rehabilitation period should be determined on the basis of the full sentence as given by the court, irrespective of any bail time.

Changed the order of the calculations for bail to reflect the correct policy.

Extended and improved some of the tests.