Remove motoring life time ban conviction
We interpreted this as there being a difference between a disqualification
period with no end date (where we applied the 2 year calculation) and a
lifetime driving ban (if users selected this, we provided them with the
information that their offence would never be spent).

We have now checked this with policy and legal, who have confirmed
disqualification period with no end date and lifetime driving ban are the same
thing, and a 2 year rehabilitation period applies.

Story: https://mojdigital.teamwork.com/#/tasks/20951241