Ability to freeze datetime on cukes
Some scenarios are time-dependent, and if we don't freeze the datetime, when time passes they may break.

With the new step `Given The current date is DD-MM-YYYY` we can set the date to whatever we want to have a predictable outcome.

Note: I've only added this new step in a couple places but there are more, we need to do the same in other places.