Update page layout to simplify markup
In CAS we managed to simplify each of the individual views with the absolute minimum necessary markup , lifting as much of it to the layout itself.

This also place the back link right where it needs to go, above the `main` container.

This commit updates the layout and govuk template to support the changes in the next commit.