Update all views (except ones having forms)
This looks like a huge commit but it was just removing the outter `div` and `main` containers from the markup in each of the views that does not contain a form.

Only one with a form that I've updated is the one that has been already migrated to use the new gem, `steps/check/kind/edit.html.erb`

The rest of form views will get their markup updated as we go. There will be an extra padding in these views but is not the end of the world, we can go one by one.