Migrate `conviction_known_date` to use new gem
Similar to previous PR to migrate gems, but this one is a bit special as this is the first form to be migrated with our custom `i18n_attribute` that we created to allow a same form to be reused in multiple views with different legends and hint texts.

This is a preliminary version of a potential solution that more or less maintain the same interface and same locales and should be easy to migrate the rest of the forms.