Migrate `motoring_disqualification_end_date` to use new gem
Ticket: https://mojdigital.teamwork.com/#/tasks/21737618

Same as with previous dates.
This one however allows to be in the future, and also to be left blank.

In a follow-up PR, I will remove all the code related to the old dates gem and refactor the duplication in the tests now that we've migrated all the dates.