Fix BaseMultiplesCalculator to consider spent_simple
Back in pull request #364 the simple caution was simplified and the date question was skipped.

While this is currently working, this work is necessary in preparation for multiple convictions.

This fixes the following error:

```
NoMethodError: undefined method `past?' for :spent_simple:Symbol
```

When a user has a simple caution, that caution is spent on the day it was given, therefore the return value is true.

Story: https://trello.com/c/Gbnpo8Hu