Update calculators to support `indefinite` length
This only affects (currently) the AdditionCalculator, as other convictions do not require the length.

Essentially, the `expiry_date` will return a magic keyword `indefinite` so we can work out in the results what to render (next commit).

The multiples calculator for same proceedings will need to be updated once we know how `indefinite` affect these.