Add new results page for indefinite convictions
If a conviction has an indefinite / until further order length, then the results page "banner" is slighly different to a "never spent", as it means, currently it is not spent, and there is no date when it will be spent, but another order can be made to change it.

Add a couple of feature tests covering this new indefinite option. No need to add more as all convictions with indefinite length behave the same.