Remove motoring Fixed Penalty Notice
A decision has been made to remove FPN from the service as these notices 
are spent on the day that they are given.
This commit removes all the references to either Youth or Adult FPN 
convictions.

All done in one commit to be easier to revert (if ever needed).

Story: https://trello.com/c/Vcpqhlqr