Refactor motoring disqualification calculator
Ticket: https://trello.com/c/PhgvFgKM

This is some preliminary work to simplify the implementation of the `indefinite` length option, and also to reduce the repetition of code.

There are no changes to the rules or the implementation of the calculators, just a refactor to make them more reusable (and look more similar to other calculators like the SentenceCalculator).

In a follow-up PR, we will change the decision tree for motoring disqualifications to ask for the ban length (instead of the ban end date) and update the calculators to work with the length.