Create specs based on given examples
This commit brings a new set of tests that have been based on some scenarios,
so that we can improve the calculatios of multiple convictions (same proceedings or separte proceedings)

I've marked a test as pending for same proceedings since we haven't tackled yet relevant orders.

Although another test showed that we are marking all separate convictions as
never spent if there's a never spent conviction in one of the groups.

This is not correct, only the convictions that overlap would be considered never spent,
any convictions which are spent before should not be considered "never spent".

Given that another conviction is added after the "never spent" conviction,
that new conviction will become never spent too.

Story: https://trello.com/c/W2K1PV43/40-dc-calculations-for-separate-proceedings