Tweak the sorting of proceedings
We only need to sort the proceedings by `start date` in the `#spent_date_for` loop, as we need to see if some drag-through is needed.

But in the "check your answers" and final results page, the order should be that of the user, so it is predictable and the user as they enter cautions or convictions, see them being added to the bottom and not jumping around depending on start date.